backend/requirements.txt
Flask
Flask-Cors
firebase-admin
google-generativeai
python-dotenv 
PyPDF2
python-docx 
Pillow 
# pytesseract 
gunicorn

